requirements.txt
canvasapi
python-dateutil
PyQt6
pywebcopy
pandas
beautifulsoup4
ics
gcsa
piazza-api
dacite

downloader @ git+https://github.com/quintenroets/downloader
libs @ git+https://github.com/quintenroets/libs

